made requirement changes, Nav bar me, removed the call and email options since there is already contact us page on the website. and added a whatsapp options kyoki usme dono call and whatsapp wala number aa jata hai
Images i've removed, will try and see if there is something that can be done.

Updated blogs to self help tools

Added life diseases and combined Lack of inner peace & finding the purpose of life in specialisation

Contact us section did not made it horizontal, because it won't be looking good on phones or tablets.